pyproject.toml
[tool.poetry]
name = "discord-music-bot"
version = "0.1.0"
description = ""
authors = ["hemanta212 <sharmahemanta.212@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.20.0"
youtube_dl = "^2021.12.17"
"discord.py" = {extras = ["voice"], version = "^2.0.0"}
Flask = "^2.2.2"
yt-dlp = "^2023.3.4"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.7.0"
pylint = "^2.7.2"
jedi = "^0.18.0"
ptpython = "^3.0.16"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
